format. TODO: Format is for pretty printing, so it's a bit ...
difficult sometimes

for example, when there are too many nested things, it will give up
indenting them:
```
asd(
  larg,
  asd(
    larg,
    asd(
     larg,
     asd(
     larg,
     asd(
```
this makes it neater but maybe you might want the
indentation to continue regardless of the line length.

idk if it has this flexibility.

but idk, it's hard ig when you have very big expressions.